+++ app/assets/javascripts/discourse/app/models/user.js
@@ -333,11 +333,9 @@ const User = RestModel.extend({
       return [];
     }
 
-    return Site.current()
-      .categoriesList.filter((category) =>
-        sidebarCategoryIds.includes(category.id)
-      )
-      .sort((a, b) => a.name.localeCompare(b.name));
+    return Site.current().categoriesList.filter((category) =>
+      sidebarCategoryIds.includes(category.id)
+    );
   },
 
   changeUsername(new_username) {
